backend/requirements.txt
# Al-Mudeer Backend Dependencies
# Zero-cost, production-ready stack
# P3-8 FIX: All versions pinned for production stability and reproducibility

# Web Framework
fastapi>=0.115.0
pytz>=2023.3
# uvicorn[standard]>=0.30.0
uvicorn>=0.30.0
httptools==0.6.1
websockets>=12.0
watchfiles>=0.21.0

# Database
aiosqlite==0.19.0
aiofiles>=23.2.1


# For local LLM (Ollama) or free API alternatives
httpx>=0.26.0

# For diagnostic scripts
aiohttp>=3.9.0


# Gmail OAuth 2.0 and API
google-auth>=2.30.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.108.0

# Validation & Settings
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
email-validator==2.1.0

# QR Code Generation
qrcode[pil]==7.4.2
pillow==10.2.0

# Security
python-jose[cryptography]>=3.4.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
cryptography>=43.0.0

# CORS
python-multipart>=0.0.22

# Performance & Caching
cachetools==5.3.2

# JSON Repair (for truncated LLM responses)
json-repair>=0.30.0
slowapi==0.1.9
redis==5.0.1

# Database (PostgreSQL support)
asyncpg==0.29.0

# Database migrations
alembic==1.13.1

# Telegram Client API (for phone number integration)
telethon==1.34.0

# Web Push Notifications
pywebpush==2.1.0
py-vapid==1.9.1

# Testing (dev dependencies)
pytest==7.4.4
pytest-asyncio==0.23.3
ruff==0.1.9

# Telethon Optimization
# cryptg (Removed for stability)



# Text-to-Speech (Google Cloud TTS via Service Account)
PyJWT>=2.8.0
# cryptography is already included above

# Date Conversion
hijri-converter>=1.3.4

# Web Scraping & Link Previews
beautifulsoup4==4.12.3
bleach==6.1.0

# File Security (Optional but recommended for production)
# Issue #27: File content validation
# Note: Use python-magic for Linux/Docker, python-magic-bin for Windows local dev
python-magic==0.4.27

# Security utilities
# Issue #28: Path traversal protection (secure_filename)
werkzeug>=3.1.6

# System Metrics
psutil>=5.9.0

# Monitoring & Observability
# Prometheus metrics for library operations and system health
prometheus-client>=0.19.0

# OpenTelemetry for distributed tracing (P5)
# Using compatible 0.54b0 versions across all instrumentation packages
opentelemetry-api==1.32.0
opentelemetry-sdk==1.32.0
opentelemetry-instrumentation-fastapi==0.53b0
opentelemetry-exporter-otlp==1.32.0
opentelemetry-instrumentation-asyncio==0.53b0
opentelemetry-instrumentation-redis==0.53b0
opentelemetry-instrumentation-psycopg2==0.53b0

# Delta Updates (bsdiff algorithm for efficient APK patches)
# Generates binary patches between APK versions (saves 60-80% bandwidth)
bsdiff4>=1.2.6
